Initial support for local notifications!
